+++ src/memcount.c
@@ -1437,6 +1437,6 @@ void report_memory_usage(aClient *cptr, int detail)
 #endif  /* MEMTRACE */
 
     sendto_one(cptr, "%ssbrk(0)-etext: %u", pfxbuf,
-               (u_int) sbrk((size_t) 0) - (u_int) sbrk0);    
+               (u_long) sbrk((size_t) 0) - (u_long) sbrk0);    
 }
